Deployer.Instance (deployer v0.1.0)

Instance is a module to hold the structure of a ec2 instance. It also handles interaction with the ec2 instance.

These interactions include: connecting via ssh, uploading files, and running commands.

Functions

connect_ssh(ip, username, private_key_file, known_hosts_file, retries \\ 1)

Specs

connect_ssh(charlist(), charlist(), String.t(), String.t(), pos_integer()) ::
  {:ok, map()} | {:error, any()}

Establish ssh connection to instance.

Parameters

* ip - IP of host to connect to
* username - User to use for login
* private_key_file - Path to private key file to use (private key of key_pair)
* known_hosts_file - Path to known hosts file
* retries - Number of retries to use on connection

map(data)

Returns list of VPC structs from xml formatted original

Parameters

* data - XML string of security groups object

run_command(connection, cmd)

Specs

run_command(map(), String.t()) :: :ok | {:error, any()}

Run command on established ssh connection.

Parameters

* connection - SSH connection from `connect_ssh`
* cmd - Command to run on instance

run_commands(connection, list)

Specs

run_commands(map(), [String.t()]) :: :ok | {:error, any()}

Run multiple commands on established ssh connection

Parameters

* connection - SSH connection from `connect_ssh`
* commands - List of commands to run

upload_file(ip, username, private_key_file, known_hosts_file, map)

Specs

upload_file(charlist(), charlist(), String.t(), String.t(), map()) :: :ok

Upload file to instance.

Parameters

* ip - IP of host to connect to
* username - User to use for login
* private_key_file - Path to private key file to use (private key of key_pair)
* known_hosts_file - Path to known hosts file
* file - Map defining `src` and `dest` of file to upload

upload_files(ip, username, private_key_file, known_hosts_file, list)

Specs

upload_files(charlist(), charlist(), String.t(), String.t(), [map()]) ::
  :ok | any()

Upload multiple files to instance.

Parameters

* ip - IP of host to connect to
* username - User to use for login
* private_key_file - Path to private key file to use (private key of key_pair)
* known_hosts_file - Path to known hosts file
* files - List of files to upload defined by maps with `src` and `dest` of file to upload