Deployer.Instances (deployer v0.1.0)

Main module for interacting with AWS.

These interactions include: preflight checks, launching instances, security group interaction, subnet interaction, and more.

Functions

bootstrap(manifest, instance, private_key_path, known_hosts_path)

Specs

bootstrap(Deployer.Manifest, Deployer.Instance, String.t(), String.t()) ::
  :ok | :error

Bootstrap instance by installing dependencies and desired files in manifest.

Parameters

* manifest - Manifest to use to deploy
* instance - Instance that has already been deployed
* private_key_path - Path to private key file to use (private key of key_pair)
* known_hosts_path - Path to known hosts file

get_instance_status(instance)

Specs

get_instance_status(Deployer.Instance) :: Deployer.Instance | :error

Get instance status.

Parameters

* instance - Instance from aws

get_security_group(vpc_id)

Specs

get_security_group(String.t()) ::
  {:ok, Deployer.SecurityGroup} | {:error, any()}

Get security group selection from user by querying EC2 and asking for choice.

Parameters

* vpc_id - VPC id already generated ahead of time

get_subnet(vpc_id)

Specs

get_subnet(String.t()) :: {:ok, Deployer.Subnet} | {:error, any()}

Get subnet selection from user by querying EC2 and asking for choice.

Parameters

* vpc_id - VPC id already generated ahead of time

get_vpc(vpc_id)

Specs

get_vpc(String.t()) :: {:ok, map()} | {:error, any()}

Checks if given vpc exists, Deployer doesn't current support creating vpcs

Parameters

* vpc_id - VPC id already generated ahead of time

launch(manifest, vpc_id, key_pair)

Specs

launch(Deployer.Manifest, String.t(), String.t()) ::
  {:ok, Deployer.Instance} | :error

Launch ec2 instance after getting user input.

Parameters

* manifest - Manifest to use to deploy
* vpc_id - Id of vpc to launch in
* key_pair - EC2 key pair to use (must already be loaded in aws)

preflight(vpc_id)

Specs

preflight(String.t()) ::
  {:ok, Deployer.Subnet, Deployer.SecurityGroup} | {:error, any()}

Run preflight to get user input for desired subnet and security group.

Parameters

* vpc_id - Id of vpc to launch in

run_instance(manifest, subnet, security_group, key_pair)

Specs

run_instance(
  Deployer.Manifest,
  Deployer.Subnet,
  Deployer.SecurityGroup,
  String.t()
) :: {:ok, Deployer.Instance} | {:error, any()}

Run instance defined in manifest.

Parameters

* manifest - Manifest to use to deploy
* subnet - Subnet to launch in
* security_group - Security group to add to instance
* key_pair - EC2 key pair to use

start_app(manifest, instance, private_key_path, known_hosts_path)

Specs

start_app(Deployer.Manifest, Deployer.Instance, String.t(), String.t()) ::
  :ok | :error

Start application that has been uploaded.

Parameters

* manifest - Manifest to use to deploy
* instance - Instance that has already been deployed
* private_key_path - Path to private key file to use (private key of key_pair)
* known_hosts_path - Path to known hosts file

wait_until_running(instance)

Specs

wait_until_running(Deployer.Instance) :: Deployer.Instance

Wait for instance to enter the running state.

Parameters

* instance - Instance from aws