Deployer.SecurityGroup (deployer v0.1.0)

SecurityGroup is a module to hold the structure of a security group. It also handles interaction with the use.

It is capable of adding ingress rules on demand for a provided security group.

Functions

add_required_rules(security_group)

Loads security_group with all required rules. Intended to be idempotent and not error if rules already exist

Parameters

* security_group - `Deployer.SecurityGroup` to add rules to

Returns :ok if no errors occured. Otherwise {:error, ERROR}

get_user_choice(security_groups \\ [])

Returns user picked security group id. Will create security group if necassary

Parameters

* security_groups - List of security group structs

map(data)

Returns list of SecurityGroup structs from xml formatted original

Parameters

* data - XML string of security groups object

user_select(bad_choice, choices, security_groups)

Returns selected security group struct based on users choice

Parameters

* choice - Tuple choice from integer check
* choices - Mapped choices to user choice
* security_groups - All existing security groups